Add dbt model contracts to core and mart layers
Contracts (contract: enforced: true) applied to all 5 public models:
  dim_date, dim_listing, dim_agent, fct_leads, mart_leads_per_listing

Each yml file now declares an exact data_type per column, giving
downstream consumers a stable, versioned interface guarantee.
dbt will fail the build if any column name, type, or count diverges
from the declared contract — verified both a name-mismatch and a
type-mismatch produce clear error tables.

SQL changes to produce clean contract-friendly types:
  - dim_agent: CASE/SUM results cast ::bigint (was HUGEINT)
  - mart_leads_per_listing: SUM/CASE results cast ::bigint,
    ROUND results cast ::double (was HUGEINT / implicit DOUBLE)

Contracts intentionally NOT applied to staging models: the staging
layer is internal and should absorb upstream schema drift gracefully.

+++ models/core/dim_agent.sql
@@ -16,8 +16,8 @@ agent_stats as (
     select
         agent_id,
         count(listing_id)                                              as total_listings,
-        sum(case when is_active then 1 else 0 end)                     as active_listings,
-        sum(case when not is_active then 1 else 0 end)                 as inactive_listings,
+        sum(case when is_active     then 1 else 0 end)::bigint          as active_listings,
+        sum(case when not is_active then 1 else 0 end)::bigint         as inactive_listings,
         min(created_at)                                                as first_listing_at,
         max(created_at)                                                as latest_listing_at,
         -- Regions the agent operates in (aggregated for reference)
